Renamed LoginUserList* to LoginAuthor*
