+++ kview/src/main/java/module-info.java
@@ -147,8 +147,8 @@ module dev.ikm.komet.kview {
 
     exports dev.ikm.komet.kview.mvvm.view.confirmation;
     opens dev.ikm.komet.kview.mvvm.view.confirmation;
-    exports dev.ikm.komet.kview.mvvm.view.loginuserlist;
-    opens dev.ikm.komet.kview.mvvm.view.loginuserlist;
+    exports dev.ikm.komet.kview.mvvm.view.loginauthor;
+    opens dev.ikm.komet.kview.mvvm.view.loginauthor;
 
     provides dev.ikm.komet.framework.KometNodeFactory with dev.ikm.komet.kview.mvvm.view.details.DetailsNodeFactory, dev.ikm.komet.kview.mvvm.view.properties.PropertiesNodeFactory;
